requirements.txt
asgiref==3.8.1
automium==0.2.6
automium_web==0.1.1
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.7
defusedxml==0.8.0rc2
Django==5.0.6
django-allauth==0.62.1
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-crispy-forms==1.14.0
django-filter==24.2
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-modeltranslation==0.18.13
django-phonenumber-field==7.3.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
Flask==3.0.3
idna==3.7
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
legacy==0.1.7
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.0
phonenumbers==8.13.37
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
setuptools==74.0.0
sqlparse==0.5.0
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.3
whitenoise==6.6.0
